requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<26.0
drf-spectacular>=0.27,<0.29
django-environ>=0.11,<0.12
psycopg[binary]>=3.1,<4.0
django-jazzmin>=3.0,<4.0
Pillow>=10.0,<12.0
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0
gunicorn>=22.0,<24.0
whitenoise>=6.6,<7.0
